Auslagerung der Verarbeitung der Detail-Infos
