+++ src/main/java/mServer/crawler/sender/arte/ArteJsonObjectToDatenFilmCallable.java
@@ -41,20 +41,10 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
     private static final String ARTE_VIDEO_INFORMATION_URL_PATTERN = "https://api.arte.tv/api/player/v1/config/%s/%s?platform=ARTE_NEXT";
     private static final String ARTE_VIDEO_INFORMATION_URL_PATTERN_2 = "https://api.arte.tv/api/opa/v3/programs/%s/%s"; // Für broadcastBeginRounded
     private static final String JSON_ELEMENT_KEY_SHORT_DESCRIPTION = "shortDescription";
-    private static final String JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_1 = "programs";
-    private static final String JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_2 = "broadcastProgrammings";
-    private static final String JSON_ELEMENT_BROADCAST = "broadcastBeginRounded";
-    private static final String JSON_ELEMENT_BROADCASTTYPE = "broadcastType";
-    private static final String JSON_ELEMENT_BROADCAST_CATCHUPRIGHTS_BEGIN = "catchupRightsBegin";
-    private static final String JSON_ELEMENT_BROADCAST_CATCHUPRIGHTS_END = "catchupRightsEnd";
-    private static final String BROADCASTTTYPE_FIRST = "FIRST_BROADCAST";
-    private static final String BROADCASTTTYPE_MINOR_RE = "MINOR_REBROADCAST";
-    private static final String BROADCASTTTYPE_MAJOR_RE = "MAJOR_REBROADCAST";
     
     private final JsonObject jsonObject;
     private final String langCode;
     private final String senderName;
-    private final Calendar today;
     
     private final FastDateFormat broadcastDateFormat = FastDateFormat.getInstance("yyyy-MM-dd'T'HH:mm:ssX");//2016-10-29T16:15:00Z
     
@@ -63,7 +53,6 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
         langCode = aLangCode;
         senderName = aSenderName;
         
-        today = Calendar.getInstance();
     }
 
     @Override
@@ -83,7 +72,10 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
             String programId = getElementValue(jsonObject, JSON_ELEMENT_KEY_PROGRAM_ID);
             String videosUrl = String.format(ARTE_VIDEO_INFORMATION_URL_PATTERN, langCode, programId);
 
-            Gson gson = new GsonBuilder().registerTypeAdapter(ArteVideoDTO.class, new ArteVideoDeserializer()).create();
+            Gson gson = new GsonBuilder()
+                    .registerTypeAdapter(ArteVideoDTO.class, new ArteVideoDeserializer())
+                    .registerTypeAdapter(ArteVideoDetailsDTO.class, new ArteVideoDetailsDeserializer())
+                    .create();
 
             try(Response responseVideoDetails = executeRequest(videosUrl))
             {
@@ -96,9 +88,9 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
 
                     if (video.getVideoUrls().containsKey(Qualities.NORMAL))
                     {
-                        String broadcastBegin = getBroadcastBegin(gson, programId);                      
+                        ArteVideoDetailsDTO details = getVideoDetails(gson, programId);                      
 
-                        film = createFilm(thema, urlWeb, titel, video, broadcastBegin, durationAsTime, beschreibung);
+                        film = createFilm(thema, urlWeb, titel, video, details.getBroadcastBegin(), durationAsTime, beschreibung);
                    } else {
                        LOG.debug(String.format("Keine \"normale\" Video URL für den Film \"%s\" mit der URL \"%s\". Video Details URL:\"%s\" ", titel, urlWeb, videosUrl));
                    }
@@ -116,8 +108,8 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
         return film;
     }
     
-    private String getBroadcastBegin(Gson gson, String programId) throws IOException {
-        String broadcastBegin = "";
+    private ArteVideoDetailsDTO getVideoDetails(Gson gson, String programId) throws IOException {
+        ArteVideoDetailsDTO details = null;
         
         OkHttpClient httpClient = MVHttpClient.getInstance().getHttpClient();
         //https://api.arte.tv/api/opa/v3/programs/[language:de/fr]/[programId]
@@ -130,129 +122,10 @@ public class ArteJsonObjectToDatenFilmCallable implements Callable<DatenFilm>
         {
           if(responseVideoDetails2.isSuccessful())
           {
-            /*
-             * Grobe Struktur des Json's:
-             * {
-             *  "meta": {}
-             *  "programs": [{                   (JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_1)
-                  "broadcastProgrammings": [{    (JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_2)
-                    "broadcastBeginRounded": "2016-07-06T02:40:00Z",     (JSON_ELEMENT_BROADCAST)
-                    "broadcastType": MAJOR_REBROADCAST, (JSON_ELEMENT_BROADCAST_TYPE)
-                  ]}
-                ]}
-             * }
-             */
-            JsonObject jsonObjectVideoDetails2 = gson.fromJson(responseVideoDetails2.body().string(), JsonObject.class);
-            if(jsonObjectVideoDetails2.isJsonObject() && 
-                jsonObjectVideoDetails2.get(JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_1).getAsJsonArray().size() > 0) {
-                
-                JsonObject programElement = jsonObjectVideoDetails2
-                    .get(JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_1).getAsJsonArray().get(0).getAsJsonObject();
-                JsonArray broadcastArray = programElement.get(JSON_ELEMENT_BROADCAST_ELTERNKNOTEN_2).getAsJsonArray();
-                
-                if(broadcastArray.size() > 0) {
-                    String broadcastBeginFirst = "";
-                    String broadcastBeginMajor = "";
-                    String broadcastBeginMinor = "";
-                    
-                    // nach Priorität der BroadcastTypen den relevanten Eintrag suchen
-                    // FIRST_BROADCAST => MAJOR_REBROADCAST => MINOR_REBROADCAST
-                    // dabei die "aktuellste" Ausstrahlung verwenden
-                    for(int i = 0; i < broadcastArray.size(); i++) {
-                        JsonObject broadcastObject = broadcastArray.get(i).getAsJsonObject();
-                        if(broadcastObject.has(JSON_ELEMENT_BROADCASTTYPE) && 
-                                broadcastObject.has(JSON_ELEMENT_BROADCAST)) {
-                            String value = this.getBroadcastDate(broadcastObject);
-                            if(!value.isEmpty()) {
-                                String type = broadcastObject.get(JSON_ELEMENT_BROADCASTTYPE).getAsString();
-                                switch(type) {
-                                    case BROADCASTTTYPE_FIRST:
-                                        broadcastBeginFirst = value;
-                                        break;
-                                    case BROADCASTTTYPE_MAJOR_RE:
-                                        broadcastBeginMajor = value;
-                                        break;
-                                    case BROADCASTTTYPE_MINOR_RE:
-                                        broadcastBeginMinor = value;
-                                        break;
-                                    default:
-                                        LOG.debug("New broadcasttype: " + type);
-                                }
-                            }
-                        }
-                    }
-                    
-                    if(!broadcastBeginFirst.isEmpty()) {
-                        broadcastBegin = broadcastBeginFirst;
-                    } else if(!broadcastBeginMajor.isEmpty()) {
-                        broadcastBegin = broadcastBeginMajor;
-                    } else if(!broadcastBeginMinor.isEmpty()) {
-                        broadcastBegin = broadcastBeginMinor;
-                    }
-                } else {
-                    // keine Ausstrahlungen verfügbar => catchupRightsBegin verwenden
-                    JsonElement elementBegin = programElement.get(JSON_ELEMENT_BROADCAST_CATCHUPRIGHTS_BEGIN);
-                    if(!elementBegin.isJsonNull()) {
-                        broadcastBegin = elementBegin.getAsString();
-                    }
-                }
-            }
+            details = gson.fromJson(responseVideoDetails2.body().string(), ArteVideoDetailsDTO.class);
           }
         }      
-        return broadcastBegin;
-    }
-    
-    /**
-     * Liefert den Beginn der Ausstrahlung, 
-     * wenn 
-     *  - heute im Zeitraum von CatchUpRights liegt 
-     *  - oder heute vor dem Zeitraum liegt
-     *  - oder CatchUpRights nicht gesetzt ist und die Ausstrahlung in der Vergangenheit liegt
-     * @param broadcastObject 
-     * @return der Beginn der Ausstrahlung oder ""
-     */
-    private String getBroadcastDate(JsonObject broadcastObject) {
-        String broadcastDate = "";
-        
-        JsonElement elementBegin = broadcastObject.get(JSON_ELEMENT_BROADCAST_CATCHUPRIGHTS_BEGIN);
-        JsonElement elementEnd = broadcastObject.get(JSON_ELEMENT_BROADCAST_CATCHUPRIGHTS_END);
-        if (!elementBegin.isJsonNull() && !elementEnd.isJsonNull()) {
-            String begin = elementBegin.getAsString();
-            String end = elementEnd.getAsString();
-
-            try {
-                Calendar beginDate = Calendar.getInstance();
-                beginDate.setTime(broadcastDateFormat.parse(begin));
-                Calendar endDate = Calendar.getInstance();
-                endDate.setTime(broadcastDateFormat.parse(end));
-
-                if(DateWithoutTimeComparer.compare(today, beginDate) >= 0 && DateWithoutTimeComparer.compare(today, endDate) <= 0) {
-                    // wenn das heutige Datum zwischen begin und end liegt,
-                    // dann ist es die aktuelle Ausstrahlung
-                    broadcastDate = broadcastObject.get(JSON_ELEMENT_BROADCAST).getAsString();
-                } else if(DateWithoutTimeComparer.compare(today, beginDate) < 0) {
-                    // ansonsten die zukünftige verwenden
-                    broadcastDate = broadcastObject.get(JSON_ELEMENT_BROADCAST).getAsString();
-                }
-                
-            } catch (ParseException ex) {
-                LOG.debug(ex);
-            }           
-        } else {
-            String broadcast = broadcastObject.get(JSON_ELEMENT_BROADCAST).getAsString();
-            
-            try {
-                Calendar broadcastCal = Calendar.getInstance();
-                broadcastCal.setTime(broadcastDateFormat.parse(broadcast));
-                
-                if(DateWithoutTimeComparer.compare(today, broadcastCal) >= 0) {
-                    broadcastDate = broadcast;
-                }
-            } catch (ParseException ex) {
-                LOG.debug(ex);
-            }            
-        }
-        return broadcastDate;
+        return details;
     }
     
     private static String getSubject(JsonObject programObject) {
