+++ src/main/java/mServer/crawler/sender/arte/ArteVideoDetailsDTO.java
@@ -1,0 +1,16 @@
+package mServer.crawler.sender.arte;
+
+public class ArteVideoDetailsDTO {
+    private String broadcastBegin = "";
+    
+    public ArteVideoDetailsDTO() {
+    }
+    
+    public String getBroadcastBegin() {
+        return this.broadcastBegin;
+    }
+    
+    public void setBroadcastBegin(String aBroadcastBegin) {
+        this.broadcastBegin = aBroadcastBegin;
+    }
+}
